updated plotter

+++ src/cone_plotter.cpp
@@ -2,26 +2,46 @@
 #include <nodelet/nodelet.h>
 
 #include <gazebo_rad_msgs/Cone.h>
+#include <gazebo_rad_msgs/RadiationSource.h>
 
 #include <rviz_visual_tools/rviz_visual_tools.h>
+#include <visualization_msgs/Marker.h>
 
 #include <Eigen/Eigen>
+#include <mutex>
 
 namespace compton_camera_python
 {
 
   /* class Cone //{ */
-  
+
   class Cone {
-  
-    public:
-      Cone(Eigen::Isometry3d pose, double angle);
-      Eigen::Isometry3d pose;
-      double angle;
+
+  public:
+    Cone(Eigen::Isometry3d pose, double angle);
+    Eigen::Isometry3d pose;
+    double            angle;
   };
-  
-  Cone::Cone(Eigen::Isometry3d pose, double angle) : pose(pose), angle(angle) {};
-  
+
+  Cone::Cone(Eigen::Isometry3d pose, double angle) : pose(pose), angle(angle){};
+
+  //}
+
+  /* class RadiationSource //{ */
+
+  class RadiationSource {
+
+  public:
+    ros::Time                        last_update;
+    gazebo_rad_msgs::RadiationSource source_msg;
+
+    RadiationSource(gazebo_rad_msgs::RadiationSource source);
+  };
+  RadiationSource::RadiationSource(gazebo_rad_msgs::RadiationSource source) : source_msg(source) {
+
+    last_update = ros::Time::now();
+  };
+
   //}
 
   /* ConePlotter //{ */
@@ -41,11 +61,24 @@ namespace compton_camera_python
 
   private:
     ros::Subscriber subscriber_cone;
-    void callbackCone(const gazebo_rad_msgs::ConeConstPtr &msg);
+    void            callbackCone(const gazebo_rad_msgs::ConeConstPtr &msg);
+    int             max_cones_;
+    double          cone_length_;
+
     std::list<Cone> cones;
-    int max_cones_;
-    double cone_length_;
-    
+    std::mutex      mutex_cones;
+
+  private:
+    ros::Subscriber subscriber_source;
+    void            callbackSource(const gazebo_rad_msgs::RadiationSourceConstPtr &msg);
+
+    std::map<int, RadiationSource> sources;
+    std::mutex                     mutex_sources;
+
+  private:
+    ros::Timer main_timer;
+    double     main_timer_rate_;
+    void       mainTimer(const ros::TimerEvent &event);
   };
   //}
 
@@ -61,12 +94,14 @@ namespace compton_camera_python
 
     nh_.getParam("max_cones", max_cones_);
     nh_.getParam("cone_length", cone_length_);
+    nh_.getParam("main_timer_rate", main_timer_rate_);
 
     // --------------------------------------------------------------
     // |                         subscribers                        |
     // --------------------------------------------------------------
 
-    subscriber_cone = nh_.subscribe("cone_in", 1, &ConePlotter::callbackCone, this, ros::TransportHints().tcpNoDelay());
+    subscriber_cone   = nh_.subscribe("cone_in", 1, &ConePlotter::callbackCone, this, ros::TransportHints().tcpNoDelay());
+    subscriber_source = nh_.subscribe("source_in", 1, &ConePlotter::callbackSource, this, ros::TransportHints().tcpNoDelay());
 
     // --------------------------------------------------------------
     // |                         publishers                         |
@@ -84,6 +119,12 @@ namespace compton_camera_python
     // Clear messages
     visual_tools_->deleteAllMarkers();
     visual_tools_->enableBatchPublishing();
+
+    // --------------------------------------------------------------
+    // |                           timers                           |
+    // --------------------------------------------------------------
+
+    main_timer = nh_.createTimer(ros::Rate(main_timer_rate_), &ConePlotter::mainTimer, this);
 
     // | ----------------------- finish init ---------------------- |
 
@@ -105,6 +146,8 @@ namespace compton_camera_python
     if (!is_initialized)
       return;
 
+    std::scoped_lock lock(mutex_cones);
+
     ROS_INFO("[ConePlotter]: got a cone!");
 
     Eigen::Isometry3d pose1 = Eigen::Isometry3d::Identity();
@@ -120,14 +163,68 @@ namespace compton_camera_python
     cones.push_back(new_cone);
 
     if (cones.size() > uint(max_cones_)) {
-      cones.pop_front(); 
+      cones.pop_front();
     }
+  }
+
+  //}
+
+  /* callbackSource() //{ */
+
+  void ConePlotter::callbackSource(const gazebo_rad_msgs::RadiationSourceConstPtr &msg) {
+
+    if (!is_initialized)
+      return;
+
+    std::scoped_lock lock(mutex_sources);
+
+    std::map<int, RadiationSource>::iterator it;
+    it = sources.find(msg->id);
+
+    if (it == sources.end()) {
+
+      RadiationSource new_source(*msg);
+
+      sources.insert(std::pair<int, RadiationSource>(msg->id, new_source));
+
+      ROS_INFO("[ConePlotter]: registering new source [%d]", msg->id);
+
+    } else {
+
+      it->second.last_update = ros::Time::now();
+      it->second.source_msg  = *msg;
+    }
+  }
+
+  //}
+
+  /* mainTimer() //{ */
+
+  void ConePlotter::mainTimer([[maybe_unused]] const ros::TimerEvent &event) {
+
+    if (!is_initialized)
+      return;
 
     visual_tools_->deleteAllMarkers();
 
-    std::list<Cone>::iterator it;
-    for (it = cones.begin(); it != cones.end(); it++) {
-      visual_tools_->publishCone(it->pose, it->angle, rviz_visual_tools::colors::TRANSLUCENT_LIGHT, cone_length_);
+    {
+      std::scoped_lock lock(mutex_cones);
+
+      std::list<Cone>::iterator it;
+      for (it = cones.begin(); it != cones.end(); it++) {
+        visual_tools_->publishCone(it->pose, it->angle, rviz_visual_tools::colors::TRANSLUCENT_LIGHT, cone_length_);
+      }
+    }
+
+    {
+      std::scoped_lock lock(mutex_sources);
+
+      std::map<int, RadiationSource>::iterator it;
+      for (it = sources.begin(); it != sources.end(); it++) {
+        visualization_msgs::Marker new_marker;
+
+        new_marker.type = visualization_msgs::Marker::CUBE;
+      }
     }
 
     visual_tools_->trigger();
